now showing terms of use with prefix label
the text for terms of use are introduced with "Terms of Use:" prefix text.

also, fixed simple markings to place each marking structure into a separate div.  this prevents a problem that showed up where multiple marking structures would get merged into a single line of text in the ouput (with no clear separation).

address the enhancement request in issue #58