better handling of non-supported input files
now when a document is passed in that's not stix, cybox, or maec [bundle or package], a nice error message is printed to the console (via xsl:message) and a short html page is produced with the error message.

if we switch to using xslt 3.0, we can generate an actual exception, which would probably be better.

this fixes issue #69